Fixed a bunch of 'easy' warnings

+++ benchmarks/berkeleyDB/bdbHashThreaded.c
@@ -86,7 +86,7 @@ main(int argc, char *argv[])
 #endif
 
 	pthread_t * threads = malloc(num_threads * sizeof(pthread_t));
-	int i ;
+	long i ;
 	for(i = 0; i < num_threads; i++) {
 	  if ((ret = pthread_create(&(threads[i]), &attr, runThread, (void *)i)) != 0){
 	    fprintf(stderr,
@@ -114,7 +114,7 @@ main(int argc, char *argv[])
 
 
 void * runThread(void * arg) {
-  int offset = (int) arg;
+  long offset = (long) arg;
   
   pthread_mutex_lock(&mutex);
   activeThreads++;
